cmd/etgrep: update to v2 execution trace format
No support for to/from links at the moment. Use "time" as the default
sort key, which in turn allows printing the trace as we process it
rather than buffering all of the data in memory.

Allow setting "-sort=goroutine" to get the old behavior of sorting by
goroutine ID, with a blank line to show when we're starting to display
the next goroutine.

Add a "-time" flag to show events from a particular timestamp, perhaps
to get more detail (such as the stack) about an event of interest.

Plus, make stack matchers work with the v2 trace API to enable all that.

Fixes #4