cmd/grstates: decouple graph generation from display
Prepare the graph's nodes and edges separately from writing them to the
output buffer. There's still some coupling between that preparation and
the DOT format (to specify attributes like weight and penwidth), but
it's a step in a good direction.

This also gets us closer to displaying the sequence of states that a
goroutine transitions through in between points where we can see its
full stack.

And aside from https://go.dev/issue/68277, the output is deterministic.